* DataLayer/FastaWriter.cpp (WriteSequence): Cast the int64_t argument of printf to long long unsigned to be compatible with the %llu format.
git-svn-id: https://svn.bcgsc.ca/svn/personal/jsimpson/partitionAssembly/PathWalker/src@42105 ad02020f-2a26-4133-947e-1829317c1416

+++ DataLayer/FastaWriter.cpp
@@ -21,9 +21,11 @@ void FastaWriter::WriteSequence(const Sequence& seq,
 	assert(m_fileHandle != NULL);
 	if (opt::rank < 0) {
 		fprintf(m_fileHandle, ">%llu %zu %g\n%s\n",
-				id, seq.length(), multiplicity, seq.c_str());
+				(long long unsigned)id,
+				seq.length(), multiplicity, seq.c_str());
 	} else {
 		fprintf(m_fileHandle, ">%u:%llu %zu %g\n%s\n", opt::rank,
-				id, seq.length(), multiplicity, seq.c_str());
+				(long long unsigned)id,
+				seq.length(), multiplicity, seq.c_str());
 	}
 }
